move create_xblock_info and get_block_info
