+++ cms/djangoapps/contentstore/views/component.py
@@ -39,8 +39,10 @@ from ..utils import get_lms_link_for_item, get_sibling_urls, reverse_course_url
 from ..helpers import get_parent_xblock, is_unit, xblock_type_display_name
 from cms.djangoapps.contentstore.xblock_storage_handlers.helpers import (
     add_container_page_publishing_info,
-    create_xblock_info,
     load_services_for_studio,
+)
+from cms.djangoapps.contentstore.xblock_storage_handlers import (
+    create_xblock_info
 )
 
 __all__ = [
