+++ cms/djangoapps/contentstore/xblock_storage_handlers/__init__.py
@@ -5,3 +5,5 @@ from .create_xblock import *
 from .helpers import *
 from .usage_key_with_run import *
 from .handle_xblock import *
+from .create_xblock_info import *
+from .get_block_info import *
