+++ cms/djangoapps/contentstore/xblock_storage_handlers/helpers.py
@@ -96,8 +96,6 @@ from ..utils import (
     is_self_paced,
 )
 
-from .create_xblock import create_xblock
-from .usage_key_with_run import usage_key_with_run
 from ..helpers import (
     get_parent_xblock,
     import_staged_content_from_user_clipboard,
@@ -106,6 +104,8 @@ from ..helpers import (
     xblock_studio_url,
     xblock_type_display_name,
 )
+from .usage_key_with_run import usage_key_with_run
+from .create_xblock import create_xblock
 
 log = logging.getLogger(__name__)
 
@@ -116,9 +116,7 @@ NEVER = lambda x: False
 ALWAYS = lambda x: True
 
 __all__ = [
-    "create_xblock_info",
     "load_services_for_studio",
-    "get_block_info",
     "get_xblock",
     "delete_orphans",
 ]
@@ -150,23 +148,7 @@ def _is_library_component_limit_reached(usage_key):
     return total_children + 1 > settings.MAX_BLOCKS_PER_CONTENT_LIBRARY
 
 
-def modify_xblock(usage_key, request):
-    request_data = request.json
-    return _save_xblock(
-        request.user,
-        get_xblock(usage_key, request.user),
-        data=request_data.get("data"),
-        children_strings=request_data.get("children"),
-        metadata=request_data.get("metadata"),
-        nullout=request_data.get("nullout"),
-        grader_type=request_data.get("graderType"),
-        is_prereq=request_data.get("isPrereq"),
-        prereq_usage_key=request_data.get("prereqUsageKey"),
-        prereq_min_score=request_data.get("prereqMinScore"),
-        prereq_min_completion=request_data.get("prereqMinCompletion"),
-        publish=request_data.get("publish"),
-        fields=request_data.get("fields"),
-    )
+
 
 
 class StudioPermissionsService:
@@ -225,199 +207,6 @@ def _update_with_callback(xblock, user, old_metadata=None, old_content=None):
     # Update after the callback so any changes made in the callback will get persisted.
     return modulestore().update_item(xblock, user.id)
 
-
-def _save_xblock(  # lint-amnesty, pylint: disable=too-many-statements
-    user,
-    xblock,
-    data=None,
-    children_strings=None,
-    metadata=None,
-    nullout=None,
-    grader_type=None,
-    is_prereq=None,
-    prereq_usage_key=None,
-    prereq_min_score=None,
-    prereq_min_completion=None,
-    publish=None,
-    fields=None,
-):
-    """
-    Saves xblock w/ its fields. Has special processing for grader_type, publish, and nullout and Nones in metadata.
-    nullout means to truly set the field to None whereas nones in metadata mean to unset them (so they revert
-    to default).
-
-    """
-    store = modulestore()
-    # Perform all xblock changes within a (single-versioned) transaction
-    with store.bulk_operations(xblock.location.course_key):
-        # Don't allow updating an xblock and discarding changes in a single operation (unsupported by UI).
-        if publish == "discard_changes":
-            store.revert_to_published(xblock.location, user.id)
-            # Returning the same sort of result that we do for other save operations. In the future,
-            # we may want to return the full XBlockInfo.
-            return JsonResponse({"id": str(xblock.location)})
-
-        old_metadata = own_metadata(xblock)
-        old_content = xblock.get_explicitly_set_fields_by_scope(Scope.content)
-
-        if data:
-            # TODO Allow any scope.content fields not just "data" (exactly like the get below this)
-            xblock.data = data
-        else:
-            data = old_content["data"] if "data" in old_content else None
-
-        if fields:
-            for field_name in fields:
-                setattr(xblock, field_name, fields[field_name])
-
-        if children_strings is not None:
-            children = []
-            for child_string in children_strings:
-                children.append(usage_key_with_run(child_string))
-
-            # if new children have been added, remove them from their old parents
-            new_children = set(children) - set(xblock.children)
-            for new_child in new_children:
-                old_parent_location = store.get_parent_location(new_child)
-                if old_parent_location:
-                    old_parent = store.get_item(old_parent_location)
-                    old_parent.children.remove(new_child)
-                    old_parent = _update_with_callback(old_parent, user)
-                else:
-                    # the Studio UI currently doesn't present orphaned children, so assume this is an error
-                    return JsonResponse(
-                        {
-                            "error": "Invalid data, possibly caused by concurrent authors."
-                        },
-                        400,
-                    )
-
-            # make sure there are no old children that became orphans
-            # In a single-author (no-conflict) scenario, all children in the persisted list on the server should be
-            # present in the updated list.  If there are any children that have been dropped as part of this update,
-            # then that would be an error.
-            #
-            # We can be even more restrictive in a multi-author (conflict), by returning an error whenever
-            # len(old_children) > 0. However, that conflict can still be "merged" if the dropped child had been
-            # re-parented. Hence, the check for the parent in the any statement below.
-            #
-            # Note that this multi-author conflict error should not occur in modulestores (such as Split) that support
-            # atomic write transactions.  In Split, if there was another author who moved one of the "old_children"
-            # into another parent, then that child would have been deleted from this parent on the server. However,
-            # this is error could occur in modulestores (such as Draft) that do not support atomic write-transactions
-            old_children = set(xblock.children) - set(children)
-            if any(
-                store.get_parent_location(old_child) == xblock.location
-                for old_child in old_children
-            ):
-                # since children are moved as part of a single transaction, orphans should not be created
-                return JsonResponse(
-                    {"error": "Invalid data, possibly caused by concurrent authors."},
-                    400,
-                )
-
-            # set the children on the xblock
-            xblock.children = children
-
-        # also commit any metadata which might have been passed along
-        if nullout is not None or metadata is not None:
-            # the postback is not the complete metadata, as there's system metadata which is
-            # not presented to the end-user for editing. So let's use the original (existing_item) and
-            # 'apply' the submitted metadata, so we don't end up deleting system metadata.
-            if nullout is not None:
-                for metadata_key in nullout:
-                    setattr(xblock, metadata_key, None)
-
-            # update existing metadata with submitted metadata (which can be partial)
-            # IMPORTANT NOTE: if the client passed 'null' (None) for a piece of metadata that means 'remove it'. If
-            # the intent is to make it None, use the nullout field
-            if metadata is not None:
-                for metadata_key, value in metadata.items():
-                    field = xblock.fields[metadata_key]
-
-                    if value is None:
-                        field.delete_from(xblock)
-                    else:
-                        try:
-                            value = field.from_json(value)
-                        except ValueError as verr:
-                            reason = _("Invalid data")
-                            if str(verr):
-                                reason = _("Invalid data ({details})").format(
-                                    details=str(verr)
-                                )
-                            return JsonResponse({"error": reason}, 400)
-
-                        field.write_to(xblock, value)
-
-        validate_and_update_xblock_due_date(xblock)
-        # update the xblock and call any xblock callbacks
-        xblock = _update_with_callback(xblock, user, old_metadata, old_content)
-
-        # for static tabs, their containing course also records their display name
-        course = store.get_course(xblock.location.course_key)
-        if xblock.location.block_type == "static_tab":
-            # find the course's reference to this tab and update the name.
-            static_tab = CourseTabList.get_tab_by_slug(
-                course.tabs, xblock.location.name
-            )
-            # only update if changed
-            if static_tab:
-                update_tab = False
-                if static_tab["name"] != xblock.display_name:
-                    static_tab["name"] = xblock.display_name
-                    update_tab = True
-                if static_tab["course_staff_only"] != xblock.course_staff_only:
-                    static_tab["course_staff_only"] = xblock.course_staff_only
-                    update_tab = True
-                if update_tab:
-                    store.update_item(course, user.id)
-
-        result = {
-            "id": str(xblock.location),
-            "data": data,
-            "metadata": own_metadata(xblock),
-        }
-
-        if grader_type is not None:
-            result.update(
-                CourseGradingModel.update_section_grader_type(xblock, grader_type, user)
-            )
-
-        # Save gating info
-        if xblock.category == "sequential" and course.enable_subsection_gating:
-            if is_prereq is not None:
-                if is_prereq:
-                    gating_api.add_prerequisite(
-                        xblock.location.course_key, xblock.location
-                    )
-                else:
-                    gating_api.remove_prerequisite(xblock.location)
-                result["is_prereq"] = is_prereq
-
-            if prereq_usage_key is not None:
-                gating_api.set_required_content(
-                    xblock.location.course_key,
-                    xblock.location,
-                    prereq_usage_key,
-                    prereq_min_score,
-                    prereq_min_completion,
-                )
-
-        # If publish is set to 'republish' and this item is not in direct only categories and has previously been
-        # published, then this item should be republished. This is used by staff locking to ensure that changing the
-        # draft value of the staff lock will also update the published version, but only at the unit level.
-        if publish == "republish" and xblock.category not in DIRECT_ONLY_CATEGORIES:
-            if modulestore().has_published_version(xblock):
-                publish = "make_public"
-
-        # Make public after updating the xblock, in case the caller asked for both an update and a publish.
-        # Used by Bok Choy tests and by republishing of staff locks.
-        if publish == "make_public":
-            modulestore().publish(xblock.location, user.id)
-
-        # Note that children aren't being returned until we have a use case.
-        return JsonResponse(result, encoder=EdxJSONEncoder)
 
 
 @login_required
@@ -513,139 +302,6 @@ def _get_source_index(source_usage_key, source_parent):
         return source_index
     except ValueError:
         return None
-
-
-def is_source_item_in_target_parents(source_item, target_parent):
-    """
-    Returns True if source item is found in target parents otherwise False.
-
-    Arguments:
-        source_item (XBlock): Source Xblock.
-        target_parent (XBlock): Target XBlock.
-    """
-    target_ancestors = _create_xblock_ancestor_info(target_parent, is_concise=True)[
-        "ancestors"
-    ]
-    for target_ancestor in target_ancestors:
-        if str(source_item.location) == target_ancestor["id"]:
-            return True
-    return False
-
-
-def _move_item(source_usage_key, target_parent_usage_key, user, target_index=None):
-    """
-    Move an existing xblock as a child of the supplied target_parent_usage_key.
-
-    Arguments:
-        source_usage_key (BlockUsageLocator): Locator of source item.
-        target_parent_usage_key (BlockUsageLocator): Locator of target parent.
-        target_index (int): If provided, insert source item at provided index location in target_parent_usage_key item.
-
-    Returns:
-        JsonResponse: Information regarding move operation. It may contains error info if an invalid move operation
-            is performed.
-    """
-    # Get the list of all parentable component type XBlocks.
-    parent_component_types = list(
-        {
-            name
-            for name, class_ in XBlock.load_classes()
-            if getattr(class_, "has_children", False)
-        }
-        - set(DIRECT_ONLY_CATEGORIES)
-    )
-
-    store = modulestore()
-    with store.bulk_operations(source_usage_key.course_key):
-        source_item = store.get_item(source_usage_key)
-        source_parent = source_item.get_parent()
-        target_parent = store.get_item(target_parent_usage_key)
-        source_type = source_item.category
-        target_parent_type = target_parent.category
-        error = None
-
-        # Store actual/initial index of the source item. This would be sent back with response,
-        # so that with Undo operation, it would easier to move back item to it's original/old index.
-        source_index = _get_source_index(source_usage_key, source_parent)
-
-        valid_move_type = {
-            "sequential": "vertical",
-            "chapter": "sequential",
-        }
-
-        if (
-            valid_move_type.get(target_parent_type, "") != source_type
-            and target_parent_type not in parent_component_types
-        ):
-            error = _(
-                "You can not move {source_type} into {target_parent_type}."
-            ).format(
-                source_type=source_type,
-                target_parent_type=target_parent_type,
-            )
-        elif (
-            source_parent.location == target_parent.location
-            or source_item.location in target_parent.children
-        ):
-            error = _("Item is already present in target location.")
-        elif source_item.location == target_parent.location:
-            error = _("You can not move an item into itself.")
-        elif is_source_item_in_target_parents(source_item, target_parent):
-            error = _("You can not move an item into it's child.")
-        elif target_parent_type == "split_test":
-            error = _("You can not move an item directly into content experiment.")
-        elif source_index is None:
-            error = _("{source_usage_key} not found in {parent_usage_key}.").format(
-                source_usage_key=str(source_usage_key),
-                parent_usage_key=str(source_parent.location),
-            )
-        else:
-            try:
-                target_index = int(target_index) if target_index is not None else None
-                if (
-                    target_index is not None
-                    and len(target_parent.children) < target_index
-                ):
-                    error = _(
-                        "You can not move {source_usage_key} at an invalid index ({target_index})."
-                    ).format(
-                        source_usage_key=str(source_usage_key),
-                        target_index=target_index,
-                    )
-            except ValueError:
-                error = _(
-                    "You must provide target_index ({target_index}) as an integer."
-                ).format(target_index=target_index)
-        if error:
-            return JsonResponse({"error": error}, status=400)
-
-        # When target_index is provided, insert xblock at target_index position, otherwise insert at the end.
-        insert_at = (
-            target_index if target_index is not None else len(target_parent.children)
-        )
-
-        store.update_item_parent(
-            item_location=source_item.location,
-            new_parent_location=target_parent.location,
-            old_parent_location=source_parent.location,
-            insert_at=insert_at,
-            user_id=user.id,
-        )
-
-        log.info(
-            "MOVE: %s moved from %s to %s at %d index",
-            str(source_usage_key),
-            str(source_parent.location),
-            str(target_parent_usage_key),
-            insert_at,
-        )
-
-        context = {
-            "move_source_locator": str(source_usage_key),
-            "parent_locator": str(target_parent_usage_key),
-            "source_index": target_index if target_index is not None else source_index,
-        }
-        return JsonResponse(context)
 
 
 def _duplicate_block(
@@ -851,40 +507,6 @@ def get_xblock(usage_key, user):
             )
 
 
-def get_block_info(
-    xblock,
-    rewrite_static_links=True,
-    include_ancestor_info=False,
-    include_publishing_info=False,
-):
-    """
-    metadata, data, id representation of a leaf block fetcher.
-    :param usage_key: A UsageKey
-    """
-    with modulestore().bulk_operations(xblock.location.course_key):
-        data = getattr(xblock, "data", "")
-        if rewrite_static_links:
-            data = replace_static_urls(data, None, course_id=xblock.location.course_key)
-
-        # Pre-cache has changes for the entire course because we'll need it for the ancestor info
-        # Except library blocks which don't [yet] use draft/publish
-        if not isinstance(xblock.location, LibraryUsageLocator):
-            modulestore().has_changes(
-                modulestore().get_course(xblock.location.course_key, depth=None)
-            )
-
-        # Note that children aren't being returned until we have a use case.
-        xblock_info = create_xblock_info(
-            xblock,
-            data=data,
-            metadata=own_metadata(xblock),
-            include_ancestor_info=include_ancestor_info,
-        )
-        if include_publishing_info:
-            add_container_page_publishing_info(xblock, xblock_info)
-
-        return xblock_info
-
 
 def _get_gating_info(course, xblock):
     """
@@ -921,328 +543,6 @@ def _get_gating_info(course, xblock):
         if prereq:
             info["visibility_state"] = VisibilityState.gated
     return info
-
-
-@pluggable_override("OVERRIDE_CREATE_XBLOCK_INFO")
-def create_xblock_info(  # lint-amnesty, pylint: disable=too-many-statements
-    xblock,
-    data=None,
-    metadata=None,
-    include_ancestor_info=False,
-    include_child_info=False,
-    course_outline=False,
-    include_children_predicate=NEVER,
-    parent_xblock=None,
-    graders=None,
-    user=None,
-    course=None,
-    is_concise=False,
-):
-    """
-    Creates the information needed for client-side XBlockInfo.
-
-    If data or metadata are not specified, their information will not be added
-    (regardless of whether or not the xblock actually has data or metadata).
-
-    There are three optional boolean parameters:
-      include_ancestor_info - if true, ancestor info is added to the response
-      include_child_info - if true, direct child info is included in the response
-      is_concise - if true, returns the concise version of xblock info, default is false.
-      course_outline - if true, the xblock is being rendered on behalf of the course outline.
-        There are certain expensive computations that do not need to be included in this case.
-
-    In addition, an optional include_children_predicate argument can be provided to define whether or
-    not a particular xblock should have its children included.
-
-    You can customize the behavior of this function using the `OVERRIDE_CREATE_XBLOCK_INFO` pluggable override point.
-    For example:
-    >>> def create_xblock_info(default_fn, xblock, *args, **kwargs):
-    ...     xblock_info = default_fn(xblock, *args, **kwargs)
-    ...     xblock_info['icon'] = xblock.icon_override
-    ...     return xblock_info
-    """
-    is_library_block = isinstance(xblock.location, LibraryUsageLocator)
-    is_xblock_unit = is_unit(xblock, parent_xblock)
-    # this should not be calculated for Sections and Subsections on Unit page or for library blocks
-    has_changes = None
-    if (is_xblock_unit or course_outline) and not is_library_block:
-        has_changes = modulestore().has_changes(xblock)
-
-    if graders is None:
-        if not is_library_block:
-            graders = CourseGradingModel.fetch(xblock.location.course_key).graders
-        else:
-            graders = []
-
-    # Filter the graders data as needed
-    graders = _filter_entrance_exam_grader(graders)
-
-    # We need to load the course in order to retrieve user partition information.
-    # For this reason, we load the course once and re-use it when recursively loading children.
-    if course is None:
-        course = modulestore().get_course(xblock.location.course_key)
-
-    # Compute the child info first so it can be included in aggregate information for the parent
-    should_visit_children = include_child_info and (
-        course_outline and not is_xblock_unit or not course_outline
-    )
-    if should_visit_children and xblock.has_children:
-        child_info = _create_xblock_child_info(
-            xblock,
-            course_outline,
-            graders,
-            include_children_predicate=include_children_predicate,
-            user=user,
-            course=course,
-            is_concise=is_concise,
-        )
-    else:
-        child_info = None
-
-    release_date = _get_release_date(xblock, user)
-
-    if xblock.category != "course" and not is_concise:
-        visibility_state = _compute_visibility_state(
-            xblock, child_info, is_xblock_unit and has_changes, is_self_paced(course)
-        )
-    else:
-        visibility_state = None
-    published = (
-        modulestore().has_published_version(xblock) if not is_library_block else None
-    )
-    published_on = (
-        get_default_time_display(xblock.published_on)
-        if published and xblock.published_on
-        else None
-    )
-
-    # defining the default value 'True' for delete, duplicate, drag and add new child actions
-    # in xblock_actions for each xblock.
-    xblock_actions = {
-        "deletable": True,
-        "draggable": True,
-        "childAddable": True,
-        "duplicable": True,
-    }
-    explanatory_message = None
-
-    # is_entrance_exam is inherited metadata.
-    if xblock.category == "chapter" and getattr(xblock, "is_entrance_exam", None):
-        # Entrance exam section should not be deletable, draggable and not have 'New Subsection' button.
-        xblock_actions["deletable"] = xblock_actions["childAddable"] = xblock_actions[
-            "draggable"
-        ] = False
-        if parent_xblock is None:
-            parent_xblock = get_parent_xblock(xblock)
-
-        # Translators: The {pct_sign} here represents the percent sign, i.e., '%'
-        # in many languages. This is used to avoid Transifex's misinterpreting of
-        # '% o'. The percent sign is also translatable as a standalone string.
-        explanatory_message = _(
-            "Students must score {score}{pct_sign} or higher to access course materials."
-        ).format(
-            score=int(parent_xblock.entrance_exam_minimum_score_pct * 100),
-            # Translators: This is the percent sign. It will be used to represent
-            # a percent value out of 100, e.g. "58%" means "58/100".
-            pct_sign=_("%"),
-        )
-
-    xblock_info = {
-        "id": str(xblock.location),
-        "display_name": xblock.display_name_with_default,
-        "category": xblock.category,
-        "has_children": xblock.has_children,
-    }
-
-    if course is not None and PUBLIC_VIDEO_SHARE.is_enabled(xblock.location.course_key):
-        xblock_info.update(
-            {
-                "video_sharing_enabled": True,
-                "video_sharing_options": course.video_sharing_options,
-                "video_sharing_doc_url": HelpUrlExpert.the_one().url_for_token(
-                    "social_sharing"
-                ),
-            }
-        )
-
-    if xblock.category == "course":
-        discussions_config = DiscussionsConfiguration.get(course.id)
-        show_unit_level_discussions_toggle = (
-            discussions_config.enabled
-            and discussions_config.supports_in_context_discussions()
-            and discussions_config.enable_in_context
-            and discussions_config.unit_level_visibility
-        )
-        xblock_info["unit_level_discussions"] = show_unit_level_discussions_toggle
-
-    if is_concise:
-        if child_info and child_info.get("children", []):
-            xblock_info["child_info"] = child_info
-        # Groups are labelled with their internal ids, rather than with the group name. Replace id with display name.
-        group_display_name = get_split_group_display_name(xblock, course)
-        xblock_info["display_name"] = (
-            group_display_name if group_display_name else xblock_info["display_name"]
-        )
-    else:
-        user_partitions = get_user_partition_info(xblock, course=course)
-        xblock_info.update(
-            {
-                "edited_on": get_default_time_display(xblock.subtree_edited_on)
-                if xblock.subtree_edited_on
-                else None,
-                "published": published,
-                "published_on": published_on,
-                "studio_url": xblock_studio_url(xblock, parent_xblock),
-                "released_to_students": datetime.now(UTC) > xblock.start,
-                "release_date": release_date,
-                "visibility_state": visibility_state,
-                "has_explicit_staff_lock": xblock.fields[
-                    "visible_to_staff_only"
-                ].is_set_on(xblock),
-                "start": xblock.fields["start"].to_json(xblock.start),
-                "graded": xblock.graded,
-                "due_date": get_default_time_display(xblock.due),
-                "due": xblock.fields["due"].to_json(xblock.due),
-                "relative_weeks_due": xblock.relative_weeks_due,
-                "format": xblock.format,
-                "course_graders": [grader.get("type") for grader in graders],
-                "has_changes": has_changes,
-                "actions": xblock_actions,
-                "explanatory_message": explanatory_message,
-                "group_access": xblock.group_access,
-                "user_partitions": user_partitions,
-                "show_correctness": xblock.show_correctness,
-            }
-        )
-
-        if xblock.category == "sequential":
-            xblock_info.update(
-                {
-                    "hide_after_due": xblock.hide_after_due,
-                }
-            )
-        elif xblock.category in ("chapter", "course"):
-            if xblock.category == "chapter":
-                xblock_info.update(
-                    {
-                        "highlights": xblock.highlights,
-                    }
-                )
-            elif xblock.category == "course":
-                xblock_info.update(
-                    {
-                        "highlights_enabled_for_messaging": course.highlights_enabled_for_messaging,
-                    }
-                )
-            xblock_info.update(
-                {
-                    # used to be controlled by a waffle switch, now just always enabled
-                    "highlights_enabled": True,
-                    # used to be controlled by a waffle flag, now just always disabled
-                    "highlights_preview_only": False,
-                    "highlights_doc_url": HelpUrlExpert.the_one().url_for_token(
-                        "content_highlights"
-                    ),
-                }
-            )
-
-        # update xblock_info with special exam information if the feature flag is enabled
-        if settings.FEATURES.get("ENABLE_SPECIAL_EXAMS"):
-            if xblock.category == "course":
-                xblock_info.update(
-                    {
-                        "enable_proctored_exams": xblock.enable_proctored_exams,
-                        "create_zendesk_tickets": xblock.create_zendesk_tickets,
-                        "enable_timed_exams": xblock.enable_timed_exams,
-                    }
-                )
-            elif xblock.category == "sequential":
-                rules_url = settings.PROCTORING_SETTINGS.get("LINK_URLS", {}).get(
-                    "online_proctoring_rules", ""
-                )
-                supports_onboarding = does_backend_support_onboarding(
-                    course.proctoring_provider
-                )
-
-                proctoring_exam_configuration_link = None
-                if xblock.is_proctored_exam:
-                    proctoring_exam_configuration_link = (
-                        get_exam_configuration_dashboard_url(
-                            course.id, xblock_info["id"]
-                        )
-                    )
-
-                if course.proctoring_provider == "proctortrack":
-                    show_review_rules = SHOW_REVIEW_RULES_FLAG.is_enabled(
-                        xblock.location.course_key
-                    )
-                else:
-                    show_review_rules = True
-
-                xblock_info.update(
-                    {
-                        "is_proctored_exam": xblock.is_proctored_exam,
-                        "was_exam_ever_linked_with_external": _was_xblock_ever_exam_linked_with_external(
-                            course, xblock
-                        ),
-                        "online_proctoring_rules": rules_url,
-                        "is_practice_exam": xblock.is_practice_exam,
-                        "is_onboarding_exam": xblock.is_onboarding_exam,
-                        "is_time_limited": xblock.is_time_limited,
-                        "exam_review_rules": xblock.exam_review_rules,
-                        "default_time_limit_minutes": xblock.default_time_limit_minutes,
-                        "proctoring_exam_configuration_link": proctoring_exam_configuration_link,
-                        "supports_onboarding": supports_onboarding,
-                        "show_review_rules": show_review_rules,
-                    }
-                )
-
-        # Update with gating info
-        xblock_info.update(_get_gating_info(course, xblock))
-        if is_xblock_unit:
-            # if xblock is a Unit we add the discussion_enabled option
-            xblock_info["discussion_enabled"] = xblock.discussion_enabled
-        if xblock.category == "sequential":
-            # Entrance exam subsection should be hidden. in_entrance_exam is
-            # inherited metadata, all children will have it.
-            if getattr(xblock, "in_entrance_exam", False):
-                xblock_info["is_header_visible"] = False
-
-        if data is not None:
-            xblock_info["data"] = data
-        if metadata is not None:
-            xblock_info["metadata"] = metadata
-        if include_ancestor_info:
-            xblock_info["ancestor_info"] = _create_xblock_ancestor_info(
-                xblock, course_outline, include_child_info=True
-            )
-        if child_info:
-            xblock_info["child_info"] = child_info
-        if visibility_state == VisibilityState.staff_only:
-            xblock_info["ancestor_has_staff_lock"] = ancestor_has_staff_lock(
-                xblock, parent_xblock
-            )
-        else:
-            xblock_info["ancestor_has_staff_lock"] = False
-
-        if course_outline:
-            if xblock_info["has_explicit_staff_lock"]:
-                xblock_info["staff_only_message"] = True
-            elif child_info and child_info["children"]:
-                xblock_info["staff_only_message"] = all(
-                    child["staff_only_message"] for child in child_info["children"]
-                )
-            else:
-                xblock_info["staff_only_message"] = False
-
-            xblock_info[
-                "has_partition_group_components"
-            ] = has_children_visible_to_specific_partition_groups(xblock)
-        xblock_info["user_partition_info"] = get_visibility_partition_info(
-            xblock, course=course
-        )
-
-    return xblock_info
 
 
 def _was_xblock_ever_exam_linked_with_external(course, xblock):
@@ -1393,78 +693,6 @@ def _compute_visibility_state(
         return VisibilityState.ready
 
 
-def _create_xblock_ancestor_info(
-    xblock, course_outline=False, include_child_info=False, is_concise=False
-):
-    """
-    Returns information about the ancestors of an xblock. Note that the direct parent will also return
-    information about all of its children.
-    """
-    ancestors = []
-
-    def collect_ancestor_info(ancestor, include_child_info=False, is_concise=False):
-        """
-        Collect xblock info regarding the specified xblock and its ancestors.
-        """
-        if ancestor:
-            direct_children_only = lambda parent: parent == ancestor
-            ancestors.append(
-                create_xblock_info(
-                    ancestor,
-                    include_child_info=include_child_info,
-                    course_outline=course_outline,
-                    include_children_predicate=direct_children_only,
-                    is_concise=is_concise,
-                )
-            )
-            collect_ancestor_info(get_parent_xblock(ancestor), is_concise=is_concise)
-
-    collect_ancestor_info(
-        get_parent_xblock(xblock),
-        include_child_info=include_child_info,
-        is_concise=is_concise,
-    )
-    return {"ancestors": ancestors}
-
-
-def _create_xblock_child_info(
-    xblock,
-    course_outline,
-    graders,
-    include_children_predicate=NEVER,
-    user=None,
-    course=None,
-    is_concise=False,
-):
-    """
-    Returns information about the children of an xblock, as well as about the primary category
-    of xblock expected as children.
-    """
-    child_info = {}
-    child_category = xblock_primary_child_category(xblock)
-    if child_category:
-        child_info = {
-            "category": child_category,
-            "display_name": xblock_type_display_name(
-                child_category, default_display_name=child_category
-            ),
-        }
-    if xblock.has_children and include_children_predicate(xblock):
-        child_info["children"] = [
-            create_xblock_info(
-                child,
-                include_child_info=True,
-                course_outline=course_outline,
-                include_children_predicate=include_children_predicate,
-                parent_xblock=xblock,
-                graders=graders,
-                user=user,
-                course=course,
-                is_concise=is_concise,
-            )
-            for child in xblock.get_children()
-        ]
-    return child_info
 
 
 def _get_release_date(xblock, user=None):
